move the logging implementation from gregord/main.go to gregord/log.go
